Extract sub-modules from de.rs
